Add test to check search link
